update bamprep and readme

+++ readcomb/bamprep.py
@@ -5,9 +5,9 @@ bamprep.py - prepare SAM/BAM file for phase change detection
 TODO: add explicit check for samtools in userpath
 -> raise informative error if not present
 
-this script will require a path to your samtools binary - 
-this can be found by running `which samtools` in the command line
-(or, alternatively, you could symlink samtools into a bin/ folder
+this script will automatically try to detect your samtools binary by default - 
+the one it uses can be found by running `which samtools` in the command line
+(alternatively, you could symlink your preferred version of samtools into a bin/ folder
 in your project dir to keep paths clean, and then provide `./bin/samtools`
 for the path - recommended)
 """
@@ -22,24 +22,28 @@ from tqdm import tqdm
 def arg_parser():
     parser = argparse.ArgumentParser(
         description='prep SAM/BAM for phase change detection', 
-        usage='python3.5 bamprep.py [options]')
+        usage='readcomb-bamprep [options]')
 
-    parser.add_argument('-f', '--fname', required=True,
-        type=str, help='File to prepare')
-    parser.add_argument('-s', '--samtools', required=True,
-        type=str, help='Path to samtools binary')
+    parser.add_argument('-b', '--bam', required=True,
+        type=str, help='BAM file to prepare')
+    parser.add_argument('-s', '--samtools', required=False,
+        type=str, help='Path to samtools binary [optional]')
     parser.add_argument('-t', '--threads', required=False,
         type=int, help='Number of threads to use [optional]')
     parser.add_argument('--index_csi', required=False,
         action='store_true', help='Use CSI index instead of BAI [optional]')
+    parser.add_argument('--no_progress', required=False,
+        action = 'store_true', help='Skip positional sorting (faster, but disables progress \
+        bars in filtering step).')
     parser.add_argument('-o', '--outdir', required=False,
         type=str, help='Directory to write to [optional] (default: current dir)')
 
     args = parser.parse_args()
 
-    return args.fname, args.samtools, args.threads, args.index_csi, args.outdir
+    return args.bam, args.samtools, args.threads, args.index_csi, \
+        args.no_progress, args.outdir
 
-def bamprep(fname, samtools, threads, index_csi, outdir):
+def bamprep(fname, samtools, threads, index_csi, no_progress, outdir):
     """
     Prepare BAM for phase change detection with readcomb.
     Script will:
@@ -69,13 +73,26 @@ def bamprep(fname, samtools, threads, index_csi, outdir):
     else:
         raise ValueError('input file is not SAM or BAM - make sure extension is correct')
 
+    # get samtools path if not specified
+    if not samtools:
+        samtools_check_cmd = 'which samtools'
+        proc = subprocess.Popen(samtools_check_cmd.split(' '),
+                stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        out, err = proc.communicate()
+        if not out:
+            raise RuntimeError("""ERROR: samtools not found. Please
+                    either ensure samtools is in your PATH or specify a
+                    samtools binary using the --samtools argument""")
+        else:
+            samtools = out.decode('utf-8')
+
     # format args
     if outdir:
         if not outdir.endswith('/'):
             outdir += '/'
     elif not outdir:
         outdir = ''
-        print('[readcomb] no outdir specified, saving to current wd')
+        print('[readcomb] no outdir specified, saving to current working directory')
 
     # filter out non proper pair, supplementary, and secondary reads
     print('[readcomb] Filtering...') 
@@ -86,32 +103,32 @@ def bamprep(fname, samtools, threads, index_csi, outdir):
     proc = subprocess.run(filter_pos_cmd.split(' '))
     proc.check_returncode() # raises CalledProcessError if proc failed
 
-    # TODO: make positional sorting + bai creation optional - ie if no progress bars wanted
+    if not no_progress:
     # sort positionally for bai file creation
-    print('[readcomb] Positionally sorting for bai creation...')
-    sort_pos_cmd = '{s} sort -O bam '.format(s=samtools)
-    if threads:
-        sort_pos_cmd += '-@{n} '.format(n=threads)
-    sort_pos_cmd += '-o {base}.sorted.bam {base}.temp.filtered.bam'.format(base=basename)
-    proc = subprocess.run(sort_pos_cmd.split(' '))
-    proc.check_returncode()
+        print('[readcomb] Positionally sorting for bai creation...')
+        sort_pos_cmd = '{s} sort -O bam '.format(s=samtools)
+        if threads:
+            sort_pos_cmd += '-@{n} '.format(n=threads)
+        sort_pos_cmd += '-o {base}.sorted.bam {base}.temp.filtered.bam'.format(base=basename)
+        proc = subprocess.run(sort_pos_cmd.split(' '))
+        proc.check_returncode()
 
-    # create bai file
-    if index_csi:
-        idx_type = '.csi'
-    else:
-        idx_type = '.bai'
-    print('[readcomb] Creating {} index file...'.format(idx_type))
-    index_cmd = '{s} index {base}.sorted.bam {out}{base}.sorted{idx}'.format(
-            s=samtools, out=outdir, base=basename, idx=idx_type)
-    proc = subprocess.run(index_cmd.split(' '))
-    proc.check_returncode()
+        # create bai file
+        if index_csi:
+            idx_type = '.csi'
+        else:
+            idx_type = '.bai'
+        print('[readcomb] Creating {} index file...'.format(idx_type))
+        index_cmd = '{s} index {base}.sorted.bam {out}{base}.sorted{idx}'.format(
+                s=samtools, out=outdir, base=basename, idx=idx_type)
+        proc = subprocess.run(index_cmd.split(' '))
+        proc.check_returncode()
 
-    # remove positionally sorted file
-    try:
-        os.remove('{base}.sorted.bam'.format(base=basename))
-    except OSError as e:
-        raise FileNotFoundError('Unable to delete {} - {}'.format(e.filename, e.strerror))
+        # remove positionally sorted file
+        try:
+            os.remove('{base}.sorted.bam'.format(base=basename))
+        except OSError as e:
+            raise FileNotFoundError('Unable to delete {} - {}'.format(e.filename, e.strerror))
 
     # sort by name and ensure file is in BAM format
     print('[readcomb] Creating read name sorted file...')
@@ -132,10 +149,11 @@ def bamprep(fname, samtools, threads, index_csi, outdir):
 
 
 def main():
-    fname, samtools, threads, index_csi, outdir = arg_parser()
-    if not os.path.isfile(samtools):
-        raise ValueError('No samtools binary found at {}'.format(samtools))
-    bamprep(fname, samtools, threads, index_csi, outdir)
+    bam, samtools, threads, index_csi, no_progress, outdir = arg_parser()
+    if samtools:
+        if not os.path.isfile(samtools):
+            raise ValueError('ERROR: No samtools binary found at {}'.format(samtools))
+    bamprep(bam, samtools, threads, index_csi, no_progress, outdir)
     print('[readcomb] Complete.')
 
 if __name__ == '__main__':
